Update to Jekyll 4 and get minimal-mistakes via Gem
Following decoupling the site's build process from GitHub Actions, we
can now build the site using latest-and-great Jekyll, and the
minimal-mistakes theme via Gem. This reduces our Gem dependency set
significantly (no longer getting all the default Jekyll gems and
dependencies), and gives us access to niceties of Jekyll 4.

Accordingly, this addresses the issues noted previously in README; and
updates the `link` and `post_url` calls to no longer be prefixed with
`site.baseurl`.

Notably with this upgrade we now see 216 Sass deprecation warnings on
site build, however it still all seems to work. This is widely
recognized as an upstream issue per [0].

[0]: https://github.com/mmistakes/minimal-mistakes/issues/4054